refactor: `getIpcMainHandle` を追加
